requirements.txt
Flask>=1.0
line-bot-sdk>=1.5.0
schedule>=0.5.0
beautifulsoup4>=4.5.3